`evilmi-sdk-simple-jump': skip whitespace
If the cursor is placed on a group of whitespace followed by a closing item,
`evilmi-sdk-simple-jump' errors with a `scan-error', due to calling
`scan-sexps', which, after skipping whitespace, sees a prematurely terminated
expression (a close tag).

Fix that by making the function skip whitespace first, allowing it to see the
correct item character. This is consistent with `evil-jump-item'.